Fixed bug 1801 - typo in the xcf decoder, condition is always false
blaffablaffa

I'm compiling sdl_image with clang, and it gives this warning:

IMG_xcf.c:523:28: warning: use of logical '&&' with constant operand [-Wconstant-logical-operand]
   l = (Uint8)(0.2990 * ((a && 0x00FF0000) >> 16)

This is of course always false, because (a && 0x00FF0000) is '1', so if shifted gives always 0. There are a couple of other similar issues in the same file.